Change the way the ideal values are computed.
The ideal nutritional values are computed from the recommended values and doctor's prescriptions.
The recommended values are computed based on the documentul facut de Dalma.
The NutrientsValueHelper changes the source from the database to the NutritionalInformation which does
all the work.
Some tests for NutritionalInformation also submitted.
A staruml document describing the flow is also submitted.